Release and removed logging statements...
Receiving shooter is working pretty well.  It is skipping some
shots, I think because it doesn't have enough time to complete
the animation before it is supposed to fire the next shot.
Should probably let the animation finish, as the exact time is less
important than finishing the animation and firing the shot.